Use separate config class for Hive metastore URI
This is required for connector extensions that use a different
method to locate the metastore.
