+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveClientModule.java
@@ -82,6 +82,7 @@ public class HiveClientModule
 
         binder.bind(HiveMetastoreClientFactory.class).in(Scopes.SINGLETON);
         binder.bind(HiveCluster.class).to(StaticHiveCluster.class).in(Scopes.SINGLETON);
+        configBinder(binder).bindConfig(StaticMetastoreConfig.class);
 
         binder.bind(TypeManager.class).toInstance(typeManager);
